Remove common, move to helper
