+++ helper/byte_seeker.go
@@ -1,4 +1,4 @@
-package common
+package helper
 
 import (
 	"bytes"
