+++ helper/version.go
@@ -1,4 +1,4 @@
-package common
+package helper
 
 var Version string
