+++ raw/edd_test.go
@@ -12,7 +12,7 @@ func TestEddRead(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	//dirTest := common.DirTest()
+	//dirTest := helper.DirTest()
 
 	tests := []struct {
 		name string
@@ -43,7 +43,7 @@ func TestEddWrite(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	//dirTest := common.DirTest()
+	//dirTest := helper.DirTest()
 
 	tests := []struct {
 		name string
