+++ raw/zon_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/xackery/quail/common"
+	"github.com/xackery/quail/helper"
 	"github.com/xackery/quail/pfs"
 )
 
@@ -17,7 +17,7 @@ func TestZonRead(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	dirTest := common.DirTest()
+	dirTest := helper.DirTest()
 	type args struct {
 	}
 
@@ -71,7 +71,7 @@ func TestZonWrite(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	dirTest := common.DirTest()
+	dirTest := helper.DirTest()
 	type args struct {
 	}
 
@@ -162,7 +162,7 @@ func TestZonWrite(t *testing.T) {
 
 			srcData := data
 			dstData := buf.Bytes()
-			err = common.ByteCompareTest(srcData, dstData)
+			err = helper.ByteCompareTest(srcData, dstData)
 			if err != nil {
 				t.Fatalf("%s byteCompare: %s", tt.name, err)
 			}
@@ -175,7 +175,7 @@ func TestZonWriteV4(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	dirTest := common.DirTest()
+	dirTest := helper.DirTest()
 	type args struct {
 	}
