+++ wce/eqg_test/step1_wce_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/xackery/quail/common"
+	"github.com/xackery/quail/helper"
 	"github.com/xackery/quail/pfs"
 	"github.com/xackery/quail/wce"
 )
@@ -28,7 +28,7 @@ func TestWceReadWrite(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	dirTest := common.DirTest()
+	dirTest := helper.DirTest()
 
 	start := time.Now()
 	for _, tt := range tests {
