+++ wce/eqg_test/step2_double_rewrite_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/xackery/quail/common"
+	"github.com/xackery/quail/helper"
 	"github.com/xackery/quail/pfs"
 	"github.com/xackery/quail/wce"
 )
@@ -19,7 +19,7 @@ func TestWceDoubleReadWrite(t *testing.T) {
 	if eqPath == "" {
 		t.Skip("EQ_PATH not set")
 	}
-	dirTest := common.DirTest()
+	dirTest := helper.DirTest()
 
 	for _, tt := range tests {
 		t.Run(tt.baseName, func(t *testing.T) {
